[change] Viewの分割
